Add files via upload
Imported the scripts from Cloudflare CDN in this POC. In the next phase, we will actually do everything locally through npm for more fine-grained control. 
This first phase of this project is just a proof-of-concept. 
Eventually, I would like to be able to bring in collaborative coding, similar to codespaces. Going to do some more research into what this might entail, or what will be needed to make it doable.